Add files via upload
Dear LCP-Team,

João and I have used Alistair's OASIS score in eICU. We thought it might be a good addition to the community. 
It is my first PR ever, and I am just a beginner to SQL. Although the code is running, it is computationally inefficient. On BigQuery, it took roughly 45 minutes to complete, thus commits to improve performance are highly welcome.

Best, 
Tristan